FIX: Update pcolor(mesh) to create grid edges
MPL 3.3 added a new interpolation routine to automatically create
edges for the grids when the input X/Y are the same shape as the
color data. This causes issues with map data due to the possiblity
of jumps in the coordinate arrays. This adds a modulo wrap to
the input arguments to do the interpolation in Cartopy before
sending the arguments on to Matplotlib.